Multiple Apps in DCOS && DCOS Runner

+++ bin/dcos-deploy.sh
@@ -2,6 +2,8 @@
 
 source helpers.sh
 
+TIMESTAMP_FILE=/tmp/.LAST_TIMESTAMP
+
 ENV_VARS=(
   DCOS
   TOKEN
@@ -21,6 +23,7 @@ ENV_VARS=(
   INSTRUMENTATION
   LOGS
   EXTENDED_LOGGING
+  MAILBOX_LOGGING
   AWS_ACCESS_KEY_ID
   AWS_SECRET_ACCESS_KEY
 )
@@ -35,26 +38,33 @@ done
 
 echo ">>> Beginning deployment!"
 
-echo ">>> Removing lasp-server from Marathon"
-curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X DELETE $DCOS/service/marathon/v2/apps/lasp-server > /dev/null
-sleep 2
+if [ -e "$TIMESTAMP_FILE" ]; then
+  # Last timestamp used to deploy an experiment
+  LAST_TIMESTAMP=$(cat $TIMESTAMP_FILE)
 
-echo ">>> Removing lasp-client from Marathon"
-curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X DELETE $DCOS/service/marathon/v2/apps/lasp-client > /dev/null
-sleep 2
+  echo ">>> Removing lasp-server-$LAST_TIMESTAMP from Marathon"
+  curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X DELETE $DCOS/service/marathon/v2/apps/lasp-server-$LAST_TIMESTAMP > /dev/null
+  sleep 2
 
-echo ">>> Waiting for Mesos to kill all tasks."
-wait_for_completion
+  echo ">>> Removing lasp-client-$LAST_TIMESTAMP from Marathon"
+  curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X DELETE $DCOS/service/marathon/v2/apps/lasp-client-$LAST_TIMESTAMP > /dev/null
+  sleep 2
+
+  echo ">>> Waiting for Mesos to kill all tasks."
+  wait_for_completion $LAST_TIMESTAMP
+else
+  echo ">>> No apps to be removed from Marathon"
+fi
 
 echo ">>> Configuring Lasp"
 cd /tmp
 
 # Memory of each VM.
-SERVER_MEMORY=8192.0
+SERVER_MEMORY=2048.0
 CLIENT_MEMORY=1024.0
 
 # CPU of each VM.
-SERVER_CPU=4
+SERVER_CPU=2
 CLIENT_CPU=0.5
 
 cat <<EOF > lasp-server.json
@@ -62,7 +72,7 @@ cat <<EOF > lasp-server.json
   "acceptedResourceRoles": [
     "slave_public"
   ],
-  "id": "lasp-server",
+  "id": "lasp-server-$EVAL_TIMESTAMP",
   "dependencies": [],
   "constraints": [["hostname", "UNIQUE", ""]],
   "cpus": $SERVER_CPU,
@@ -71,7 +81,7 @@ cat <<EOF > lasp-server.json
   "container": {
     "type": "DOCKER",
     "docker": {
-      "image": "vitorenesduarte/lasp-dcos",
+      "image": "vitorenesduarte/lasp-multiple-dcos",
       "network": "HOST",
       "forcePullImage": true,
       "parameters": [
@@ -99,6 +109,7 @@ cat <<EOF > lasp-server.json
     "INSTRUMENTATION": "$INSTRUMENTATION",
     "LOGS": "$LOGS",
     "EXTENDED_LOGGING": "$EXTENDED_LOGGING",
+    "MAILBOX_LOGGING": "$MAILBOX_LOGGING",
     "AWS_ACCESS_KEY_ID": "$AWS_ACCESS_KEY_ID",
     "AWS_SECRET_ACCESS_KEY": "$AWS_SECRET_ACCESS_KEY"
   },
@@ -121,7 +132,7 @@ cat <<EOF > lasp-server.json
 }
 EOF
 
-echo ">>> Adding lasp-server to Marathon"
+echo ">>> Adding lasp-server-$EVAL_TIMESTAMP to Marathon"
 curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X POST -d @lasp-server.json "$DCOS/service/marathon/v2/apps?force=true" > /dev/null
 sleep 10
 
@@ -130,7 +141,7 @@ cat <<EOF > lasp-client.json
   "acceptedResourceRoles": [
     "slave_public"
   ],
-  "id": "lasp-client",
+  "id": "lasp-client-$EVAL_TIMESTAMP",
   "dependencies": [],
   "cpus": $CLIENT_CPU,
   "mem": $CLIENT_MEMORY,
@@ -138,7 +149,7 @@ cat <<EOF > lasp-client.json
   "container": {
     "type": "DOCKER",
     "docker": {
-      "image": "vitorenesduarte/lasp-dcos",
+      "image": "vitorenesduarte/lasp-multiple-dcos",
       "network": "HOST",
       "forcePullImage": true,
       "parameters": [
@@ -166,6 +177,7 @@ cat <<EOF > lasp-client.json
     "INSTRUMENTATION": "$INSTRUMENTATION",
     "LOGS": "$LOGS",
     "EXTENDED_LOGGING": "$EXTENDED_LOGGING",
+    "MAILBOX_LOGGING": "$MAILBOX_LOGGING",
     "AWS_ACCESS_KEY_ID": "$AWS_ACCESS_KEY_ID",
     "AWS_SECRET_ACCESS_KEY": "$AWS_SECRET_ACCESS_KEY"
   },
@@ -184,6 +196,8 @@ cat <<EOF > lasp-client.json
 }
 EOF
 
-echo ">>> Adding lasp-client to Marathon"
+echo ">>> Adding lasp-client-$EVAL_TIMESTAMP to Marathon"
 curl -s -k -H "Authorization: token=$TOKEN" -H 'Content-type: application/json' -X POST -d @lasp-client.json "$DCOS/service/marathon/v2/apps?force=true" > /dev/null
 sleep 10
+
+echo $EVAL_TIMESTAMP > $TIMESTAMP_FILE
